Azure AI Agent/requirements.txt.txt
python-dotenv 
azure-identity 	
azure-a-projects
azure-ai-agents